Agregando factories y seeds de la api

+++ database/migrations/2021_04_25_172230_create_transactions_table.php
@@ -16,7 +16,7 @@ class CreateTransactionsTable extends Migration
         Schema::create('transactions', function (Blueprint $table) {
             $table->id();
             $table->integer('quantity');
-            $table->integer('buyer_id');
+            $table->unsignedBigInteger('buyer_id');
             $table->foreignId('product_id');
             $table->timestamps();
